feat(proxy-command): move proxy to ember cdv:proxy
