+++ lib/commands/index.js
@@ -6,6 +6,7 @@ module.exports = {
   'cordova:platform'                : require('./platform'),
   'cordova:plugin'                  : require('./plugin'),
   'cordova:prepare'                 : require('./prepare'),
+  'cordova:proxy'                   : require('./proxy'),
   'cordova:serve'                   : require('./serve'),
   'cordova:make-icons'              : require('./make-icons'),
   'cordova:make-splashes'           : require('./make-splashes'),
